feat: "Install PR" UI (#197)
* refactor: export pull requests selector to dedicated component

* refactor: regroup launcher+mods collapses in one collapse component

* refactor: load pull requests when opening selector collapse item

* refactor: review progress loaders' style

* fix: don't fetch PRs if they've already been loaded

* feat: update collapse style

* refactor: remove fetch success notification

* refactor: both collapses can be opened at the same time

* fix: non-accordion collapse sends an object as event parameter
